+++ infra/arima/IArima.cs
@@ -3,6 +3,6 @@ namespace uni_elastic_manager.Infra
     ///<summary>A contract for any method used to call ARIMA, which may be any of them using R.</summary>  
     public interface IArima
     {
-        int Calculate(string[] metrics);
+        double Calculate(string[] metrics);
     }
 }
